// gen/esi.ts
export namespace esi {
// ...
    export namespace character {
// ...
        export interface Contact {
            contact_id: number;
            contact_type: esi.character.ContactType;
            /**
             * Whether this contact is in the blocked list. Note a missing value denotes unknown, not true or false.
             */
            is_blocked?: boolean;
            /**
             * Whether this contact is being watched.
             */
            is_watched?: boolean;
            /**
             * Custom label of the contact.
             */
            label_id?: number;
            /**
             * Standing of the contact.
             */
            standing: number;
        }
        export interface ContactLabel {
            label_id: number;
            label_name: string;
        }
        export const enum ContactType {
            CHARACTER = "character",
            CORPORATION = "corporation",
            ALLIANCE = "alliance",
            FACTION = "faction"
        }
// ...
    }
// ...
}
// ...
export interface Responses {
// ...
    delete_characters_character_id_contacts: undefined;
    /**
     * The type of this member is the response type of for the route: [`GET /v1/characters/{character_id}/contacts/`](https://esi.evetech.net//#!/Contacts/get_characters_character_id_contacts).
     */
    get_characters_character_id_contacts: esi.character.Contact[];
    /**
     * The type of this member is the response type of for the route: [`GET /v1/characters/{character_id}/contacts/labels/`](https://esi.evetech.net//#!/Contacts/get_characters_character_id_contacts_labels).
     */
    get_characters_character_id_contacts_labels: esi.character.ContactLabel[];
    /**
     * The type of this member is the response type of for the route: [`POST /v1/characters/{character_id}/contacts/`](https://esi.evetech.net//#!/Contacts/post_characters_character_id_contacts).
     */
    post_characters_character_id_contacts: number[];
    /**
     * The type of this member is the response type of for the route: [`PUT /v1/characters/{character_id}/contacts/`](https://esi.evetech.net//#!/Contacts/put_characters_character_id_contacts).
     */
    put_characters_character_id_contacts: undefined;
// ...
}